Fix issue with invert selection on KisEncloseAndFillPainter
Previously the whole filled regions mask was inverted and then intersected
with the enclosing mask. But inverting a small mask resulted in a mask as big
as the whole image. This caused slow downs on big images when intersecting.
Now a custom invert is used that only operates on a smaller rect.