Change the fill tool to start filling on mouse down instead of up
This improves maybe a few milliseconds the perceived speed of filling. It is not
a performance improvement, but a ux one, since the filling is the same. It just
begins a bit earlier now.

A new "continuous fill mode", "do not use", is introduced. This way it is easier
to differentiate between normal fill and drag+fill modes of operation.