Make Ogg render consistently check for Theora
To decide whether to show Ogg in the combobox, Render Animation was
checking for the Dirac codec. But when checking for expected encoders,
it was checking for libtheora.

If the user's FFmpeg supported Dirac and not Theora, the list of
encoders would be empty. This would cause a crash on selecting Ogg.
To avoid this, check for the codec that's actually going to be used.


(cherry picked from commit c1a25a38fbd6b9c1550b4b2f1a4297bef7f9e2d4)