Add Android/x86 build jobs
The patch add only one manual job that can be started by the
developers explicitly. Android/x86 builds are excluded from
nigtly and release builds.

The plan is the following:

1) Make a release without Android/x86 platform
2) If the users report a regression about the missing platform,
   then we will quickly resurrect the platform.

We think that all uses of Android/x86 are currently because
of mistakenly installed APKs, downloaded from our website.
All systems, which Play Market reports using these packages,
support x86_64, so we guess that installation of x86 was a
user mistake/misclick.