[mlt] Generate proper silence for ranges without audio data
Our producer generates audio with "compressed" frequency to make sure
that it is played slower on the real device (really hacky!). When there
is no audio data, then AVformat producer generates an empty (uncompressed)
frame with 48kHz sample rate. That breaks playback speed for us.

To workaround it, we generate our custom (compressed) silence with
custom sample rate to make sure it is played slow enough.