Fix UI font size on Android
Krita attempts to modify the font size to make docker text smaller
and welcome page text larger, but it expects the size to be in points,
while Android may give it in pixels, so it only caused a warning.
This fixes it to use the pixel size, making dockers more readable
as more text can fit in the space.