Fix conflicts between mouse and touch actions requested concurrently
When drawing with a stylus on a touch-capable laptop it may happen
that a touch event will come in the middle of a stylus stroke. In
such a case we should avoid starting any touch action, while a
tablet action is still running.

This patch basically replaces all the checks for `if (!m_d->runningShortcut)`
with `if (!hasRunningShortcut())` which checks for all the types of the
actions possible. It basically prohibits running any touch action while
table action is running and vice versa.

BUG:489537