Improve KisAspectRatioLocker for KisDoubleParseUnitSpinBox
If return unit for both KisDoubleParseUnitSpinBox is not the same, then aspect ratio locker won't work properly
as set/get values don't use the same unit.

This commit use changeValuePt() and valuePt() to ensure that, whatever the return unit set for a KisDoubleParseUnitSpinBox,
they both use the same unit (Pt) to be synchronized.