Fix Python invalid escape sequence warnings
If Python finds a string with an invalid backslash escape such as '\*',
it will throw a Syntax or Deprecation warning. In the future this will
become an error. The fix is to use a raw string r'\*' instead, which
won't attempt to interpolate the escape sequences in the regexes.

BUG:489526
(cherry picked from commit 5d44af277b005692241a09f30e11bb0d16166823)