Fix bug on KisDoubleParseUnitSpinBox
When an external unit manager is deleted, call to set default unit manager back can generate a segmentation fault.
This is related to access to the "old unit manager" properties.

The fix consist to ensure that in the case the unit manager has been destroyed, there's no access anymore to it.