Reduce delay of Overlay, Histogram and Layer Thumbnails updates
It was reported that the update delay of the aforementioned previews
is too slow in Krita, so this patch tries to address the issue
by reducing the idle watcher delay to 50 ms.

It seems like these days such huge delay is not really needed, because
all the tasks are split into multithreaded jobs, which are **cancelled**
when the user actually starts doing any sane action. When the code was
originally written, Overview and Layer Thumbnails strokes were not
multithreaded, so they couldn't be cancelled with decent granularity.
Now it is not the case anymore.

The patch needs testing from real painters who use Overview docker on
a regular basis.

The patch is based on the work of Zhiqi Yao, who proposed this MR:
https://invent.kde.org/graphics/krita/-/merge_requests/2180

This patch reduces the scope of the original MR to
Overlay, Histogram and Layer Thumbnails tasks only, not changing
the idle watcher's defaults. It also adds a small sanity check that
would help us catching issues in the future.

CC:kimageshop@kde.org