Fix layer activation issue when opening multiple documents
... and switching between them.

KisNodeManager may emit node change signal right inside setCanvas()
call, which may be received by LayerBox **before** it gto received
its own setCanvas() call. This workaround just checks if the node-to-
be-activated actually belongs to the current image or not.

Ideally, we should make the active layer code a bit more straight-forward
and ensure that there is only one source of truth. But I'm not sure how
to do that right now. Perhaps, during the QML rewite?

BUG:490843