Workflow buttons: Fix icon filetype filter string
In the filetype filter for loading a custom workflow button icon,
.png had a comma after it, so it didn't work, only .svg worked.
However I see no reason to limit it to only .png and .svg,
might as well query the supported formats of QImageReader.

Also, the filter string shouldn't marked for translation,
it won't work if modified (including the position within the string).