Don't redefine Q_ASSERT when building with Qt6
To quote Arjen:

"The error has nothing to do with categorised logging, but is due to
`kis_debug.h` redefining `QT_ASSERT()`. The change that happened is that Qt 6
at some point added a `QT_ASSERT()` check to `QVector2D::operator/()` which is
marked as `inline constexpr`. This works fine with Qt's version of
`QT_ASSERT()` but your version adds logging which calls into non-constexpr
functions which is not allowed in a constexpr function. Due to the inlining,
the compiler uses whatever declaration that can be found at the callsite for
the macro.

I don't think there is a way to make categorised logging work in a constexpr
context. Personally I would recommend to not redefine macros like that and
instead use a separate macro if you want one with logging."