Don't use xsimd::default_arch in the pixel scaler code
Gentoo seems to build Krita with a global -march=znver4 flag,
which causes "scalar" version of the code to be compiled with
AVX512. Even though it is not how the system is supposed to work
(and this avx512-optimization will never be used in real runtime),
we shouldn't fail compilation.

https://bugs.gentoo.org/936296
https://github.com/xtensor-stack/xsimd/issues/1044