CPMT: Include all possible language names in the lang combobox
Because of how CPMT (ab)uses QLocale to generate the list of languages,
some of the language names end up being regional-specific (e.g.
"Simplified Chinese", "American English") and causes confusion. Without
a way to get a regional-agnostic language name, the best alternative is
to list all the regional names of the language.