Fix infinite loop with combination of clone + transform-mask-on-source
Some tricky combination of transform masks and clone layers could
cause an infinite loop, e.g. this one:

+-----------------+
|root             |
|  clone_of_p1 -+ |
|  paint 2      | |
|  paint 1    <-+ |
|    transf1      |
+-----------------+

When a clone layer is updated, it may request recalculation of the
source layer in the areas outside the bounds of the image. This
outside-the-image update could cause the async regeneration of the mask.
And async regeneration of the mask could cause an update of the clone
layer, which would close the infinite loop.

CCBUG:492320