Fix more cycling updates in clone/transform-masks combinations
If a layer has multiple transform masks (or any other masks with
complex need rects), we the transform mask update rect may leak
from outside-image position to the inside-image position, which
would cause the transform mask to start an async update. Hence we
cannot use the update rect position for prohibiting the transform
mask async updates.

To fix this issue we now have KisRenderPassFlags flags that are
passed to KisAsyncMerger and define is transform masks should do
any async updates. When a clone layer starts preparing hidden area
for itself, it just sets this flag and guarantees that no cycling
updates would happen.

BUG:443766