Add a unittest for testing clone layer updates while being hidden
The test check various combinations of clone layer updates:

* when the source is hidden
* when the clone itself is hidden
* when the clone is hidden transitively (i.e. the parent layer is hidden)

It also tests the complex cases of updates **outside** the image bounds.
The clone layer will bring this data into the visible area by the offset
it introduces.

There is a known bug currently: if the clone layer is hidden and the source
layer is updated outside the image bounds, then this change will not be
visible after making the layer visible again. These two tests are marked
with Q_EXPECT_FAIL.