Group Layer: Fix conversion of Group to Animated Layer.
Conversion of a group to a single animated layer was already working,
but users are likely to have a child layer (the one they're working on)
selected, leading to confusing behavior.

We now convert as expected if a group is selected, and
when a non-group child is selected we move up the tree
to the parent group and convert that instead.

This doesn't work with single paint layers, as by design.

BUG:476583