LayerBox: Node RClick > Convert > Convert GROUP to Animated Layer
The previous option "Convert to Aniamted Layer" seems to have been
a partially implemented feature, or placeholder or something.

1. I've switched this action with the other, more useful
"Convert Group to Animated Layer" action.

2. For regular non-group PaintLayers, I'm not sure what the value
of converting it to an animated layer is or how that' different
to just adding a keyframe. (Though I can add it if it's something
that someone really needs.)

CCBUG:476583