Fix pixel-perfect smoothing action
There was some mixup between a copy of the simple smoothing action
and the pixel-perfect smoothing action, which caused a warning.

The smoothing combobox order and enum order are expected to match
for remembering the selected option. Which means pixel must be last
unless that is refactored.

Plus some code simplification.