Fix incorrect threaded image access in multiple clone layers
We used to do an additional KisAsyncMerger pass in KisUpdateOriginalVisitor,
which could cause the walker to go out of its declared accessRect().
It could happen when the source layer had transform masks with non-
affine transformations. For such transformations, needRect depends on
the extent() of the parent layer, hence accessRect() may change between
the walker pass and the async merger pass.

Now the design is slightly changed. The source layers of the clones
are added to the main pass of the walker, so they are processed in the
same async merger and their application rect is guaranteed to stay
stable.

BUG:449964