Fix clone layers resetting the animation cache
KisLayer::updateClones() used to emit an update signal that used
to reset the animation cache. The resetting happened in a three-fold
way:

1) In LoD mode, KisSuspendProjectionUpdatesStrokeStrategy compressed the
   clone's update and issued it after the the actual stroke is finished.
   That caused the clone update to happen outside the frames lock and
   reset the cache.

2) When switching time KisRegenerateFrameStrokeStrategy used to issue
   an update under a lock, but the request leaked in LoD mode.

3) KisRecalculateTransformMaskJob::run() used to request projection
   update with the frames locked. But this lock didn't cover the updates
   issued for clones.

The actual solution of this patch is to forward the "don't invalidate frames"
flag down to the merge walkers, so that they could forward this flag to
the clones and make sure that clones don't reset frames.

TODO: we still need to properly split the walkers to avoid that nasty
      virtual inheritance and double-setting of the flags in the
      constructor.

BUG:484353