Don't build pykrita twice
It somehow happened that we ended up building pykrita plugin twice.
The static library was used in kritarunner, but the module was used
as a plugin for Krita.

Now the shared code is moved to the static library and the module
just links to this static library.


(cherry picked from commit b650ba70b5c13ce2de387548a46e30784c924dbc)

Co-authored-by: Dmitry Kazakov <dimula73@gmail.com>