Revise the color palette code
The main problem was that KoColorSet just didn't have the necessary
signaling to work as input for Qt item models. As a result more and more
"cheats" to bypass the MVC patterns were added, to the point where it just
stopped working properly altogether.

So this commit adds the bare minimum of signals to implement an item model
without dirty tricks.
The alternative would be to ensure that a KoColorSet is only ever accessed
through a single KisPaletteModel globally.

KisPaletteEditor still handles a few things that the resource system should
handle internally, but the main issue was the updatePalette() function that
was abused a lot. It was meant for editing the palette structure and nothing
else. It gets replaced by startEditing() and endEditing(). The former copies
the palette structure into staging data, which can then be edited until
endEditing() is called to apply (or discard) the changes.

Besides restoring the ability to actually edit palettes, some unknown bugs
also surfaced:
- Using the editing dialog, the resource was always renamed and thus saved,
  even if the name didn't actually change.
- The row count of new groups was ignored, it always used the default.
- The editor never disconnected signals when switching to a different
  palette.