GUI: Restored undo/redo mainToolBar actions.
It looks like these actions were unsucessfully moved
from "mainToolBar" to "editToolBar" at some point.
This was causing the undo/redo buttons to not appear
on the main toolbar anymore.

This took a bit longer to debug than I'd have hoped
due to the difficulty with bisecting and unfamiliarity
with some of the xmlgui stuff.

BUG:488452