Update version to 5.2.8 (android-only)
This release uses android-34 as target SDK and increases minimum
SDK version to android-24 (which effectively drops support of
Android 6.x). From now on, the minimum supported version of Android is
Android 7.0.

The minimum version increase was necessary for libflac was not possible
to compile with new NDK 27, because of the issues with fseeko on
armeabi-v7a.