Implement KisAbstractProjectionPlane::looseUserVisibleBounds()
The function does almost exactly the same thing that tightUserVisibleBounds()
does, but uses extent() instead of exactBounds() of the source layer as a base
for the calculation. It makes it possible to use this method in hot places, like
change/needRect calculation the transform masks.

The patch fixes two bugs:

1) needRect of a group layer is not nullified by a transform mask attached to a group.
After loading the layer, the group's original is empty, so we cannot use
node->original()->extent() as an interestRect for the non-affine transforms.

2) When calculating KisGroupLayer::extent() in pass-though mode we should use
tight/looseUserVisibleBounds() on the children to make sure the layer styles' state
is taken into account. Otherwise, changing visibility of a pass-through group containing
a layer with layer styles will cause artifacts on the canvas.

BUG:491447