Disable possibility of adding "Zoom Action" widget to the toolbars
Our current design of XMLGui and of the zoom action doesn't support having multiple
per-document actions. We have only one zoom action, which is added to the status bar.
And this action is switched manually on the document switch. We cannot just add this
action to the toolbar. It needs a lot of fixes in how KXMLGui synchronizes with
KActionCollection.

Personally, I hope this KXMLGui thing will somehow go away with the QML rewrite...

BUG:493456