[string freeze exception] Recover "Clean Up" button in the Recordings Manager
When the manager was refactored into supporting QDialogButtonBox, then clean-up
button was replaced with a "Discard" button. Even though it sounds sensible in
English, it has a different meaning in QDialogButtonBox and, therefore, in translations.

In translations, the button is usually translated as "Close without saving", which is
not what the button actually does in Krita.

BUG:455207