Switch color history in the popup palette to use last-used sorting
In the current implementation the popup palette uses hue-based sorting
for the last-used colors. This patch switches it to the last-used order.

I'm not really sure it is a good idea, since the old method seem to tried to align
the color hues with the wheel of the triangular color selector. And the new
version doesn't do that.

BUG:441900