Fix a possible saving lockout due to incorrect ownership of the saving mutex
Steps to reproduce:

1) Start a very long stroke (that would be calculated for 15 more seconds at least)
2) Press Ctrl+S to initiate saving
3) When asked to wait for the image operation to complete, press "Save without waiting"
4) When the next waiting dialog appears, press "Cancel" (you should be very quick)

Before this patch, the document will fall into a locked up state, when no saving is
possible. It happened because of a 'return' expression after the saving mutex has
been taken. This patch changes the ownership of this mutex to a unique_lock, so
returns and asserts will no longer cause a complete lockout.

This dialog could have been triggered by the autosave operation. If the used
"accidentially" clicked "Cancel" button, then the document could end up in
a locked up state as well.

BUG:496018