Make sure that point-based snap points have higher priority than line-based ones
Basically, if grid intersection lays near (or on the top of) the guide (or image bounds),
then this intersection should have higher priority than snapping to a "mere" line.

Ideally, we should make some clver resolver that can also find intersections between
the different snap strategies, e.g. inteersection between a guide and a grid line. But it
needs quite a lot of redesing in our snapping code, so we can postpone it now.

BUG:492434