Don't allow lowering/raising a mask into a locked layer
This bug could happen if a layer had a mask, and that mask
was moved down (lowered) into the locked fill layer (background).

I described this bug in this KA post: https://krita-artists.org/t/volunteers-needed-krita-bughunt-month-started/104916/35?u=yrh

Perhaps this could be improved if the mask could skip the locked node and move to yet another potential parent, but this is just a basic fix for now.

FYI @dkazakov