Fix display profile conversion flags to be updated on settigns change
Basically, our display rendering code should manipulate **not** with the color
profice, but with a triplet {profile, intent, flags}. And when comparing for the settings
change we should check all the three, not just color profile.

The patch implements KisDisplayConfig that abstracts this triplet. Now all display
rendering code manipulates with KisDisplayConfig instead of a raw KoColorProfile.

Think in the future: we might want to rename KisDisplayProfile into something like
KoColorConvertionConfig and use it for color conversions like KoColorSpace::convertTo()
instead of passing all three arguments manually.

BUG:496388