[string freeze exception] [feature] Add Unify Layers Color Space action
Sometimes the user may end up with the image consisting of layers in different color
spaces, e.g. sRGB vs Adobe RGB. That might significantly decrease the speed,
with which Krita renders the image, making it almost unusable to work with
(if the size of the image is like 50 layers in 4k in float16 space).

The patch implement a simple hot-fix for this issue. When triggered,
"Unify Layers Color Space" walks through all the layers in the document
and converts them into the image color space. It may increase the rendering speed
up to 2x times(!) in some circumstances.

The action is not present in menus (to not break xmlgui for our users). To trigger the action
use the "Search Action" shortcut:

1) Press Cltr+Enter
2) Type "Unify" to find the action
3) Confirm with Enter key

CC:kimageshop@kde.org
CC:kde-i18n-doc@kde.org