Add a "mismatched layer color space" warning icon
Sometimes the user may mix layers of different color spaces in the
image. That is a valid usecase for some users (to keep the original
quality of layer data), but it might cause significant performance
issues on huge (I mean, reeaaally huge) files. So we should warn users
about that with a not-very-critical icon (and a tooltip).

Even though the icon is technically a "warning", it should not cry
into the user's face, since it might be a well-thought usecase by
the user.

The icon can be removed from the layer only by converting the layer
into the color space of the image via:

* Ctrl+Enter: "Unify Layers Color Space"
* Layer->Convert->Convert Layer Color Space

CC:kimageshop@kde.org