Improve rendering of pattern thumbnails
Previously pattern thumbnails were rendered blurrily,
and large patterns showed more of the pattern on 1x display scaling
than on hi-dpi scaling (which is backwards).
Some patterns appeared as indistinguishable grey blobs.

Now, the pattern thumbnails are rendered more crisply,
and large patterns use the hi-dpi pixels to show more of the pattern
while keeping it at the same zoom as 1x display scaling.
The zoom is kept above 50% so that details will be distinguishable.