Fix backward compatibility of Per-Channel filter
In Krita 5.1.0 we added two additional "virtual" channels to the
per-channel filter ("Hue" and "Saturation"). And we added them **not**
to the end of the list, but **before** the "Lightness" channel, which
basically broke all the files created in Krita version before 5.1.0.

The patch manually remaps the channels depending on the number of channels
present in the configuration XML.

BUG:497336