Fix an assert when modifying Mesh Gradient on a shape
We shouldn't add any commands to the global undo stack in the mouse
move event handler. Instead, we should accumulate the command and add
it to the stack in teh very end of the operation in createCommand()
method.

The patch also fixes a memory leak, which caused the temporary command
to leak in case when the merge-branch was taken.

BUG:496519