[critical] Fixed opacity being applied twice(!) in some brushes
It seem like during the brushes port to lager in 2022, I made a
mistake in understanding what is `m_useSeparateStrengthValue` in
`KisCurveOption`. It caused `KisOpacityOption` to apply opacity twice
for some of the brushes.

It means that for two years when the user selected any opacity for
the brush, the actual opacity used for the brush was **squared**.
I.e. when the user selected opacity 0.5, the real opacity was 0.25.

This patch fixes the issue, but it **changes behavior** of existing
brush presets.

The list of affected brush engines:

* Curve
* Clone
* Deform
* Hairy
* Hatching
* Sketch
* Spray

The most popular brush engines, like Pixel and Color Smudge are **not**
affected.

I'm not sure this change should be backported to krita/5.2 since it
changes behavior.

BUG:466368
CC:kimageshop@kde.org