Qt6: Use float instead of qreal
In Qt6, a lot of qreal api has been replaces with float api:

class QColor {
...
    void getHsvF(float *h, float *s, float *v, float *a = nullptr) const;
...
}

Which gives

/home/halla/dev/krita/libs/widgets/KisHsvColorSlider.cpp: In function ‘void {anonymous}::fromQColor(QColor, QColor, HSVColor&, HSVColor&)’:
/home/halla/dev/krita/libs/widgets/KisHsvColorSlider.cpp:47:18: error: cannot convert ‘qreal*’ {aka ‘double*’} to ‘float*’
   47 |     minC.getHsvF(&min.h, &min.s, &min.v);
      |                  ^~~~~~
      |                  |
      |                  qreal* {aka double*}

Stackoverflow (...) says:

The best practice would be to eliminate the qreal usage in your codebase
regardless this case. When the contributors went through the Qt 5
refactoring, the qreal ancient concept was dropped as much as possible
and definitely should not be used much in new code where the API deals
with float.

https://stackoverflow.com/questions/27680018/error-cannot-convert-float-to-qreal-aka-double-in-initialization