Qt6 Use QColorSpace::description in prefences dialog
In Qt6, QSurfaceFormat returns a QColorSpace instead of a
QSurfaceFormat::ColorSpace. QColorSpace _has_ an enum with
named colorspaces similar to QSurfaceFormat, but no way to
check which named colorspace is current in the QColorSpace.

For now, just use the description() method to show something
in the preferences dialog box, but this is certainly not
going to be the same thing.