Fix KoColorSet::loadAct
The loadAct code would read beyond the end of a byte array. In Qt5
this only gives a warning:

QWARN  : TestKoColorSet::testLoadACT() Using QByteRef with an index pointing outside the valid range of a QByteArray. The corresponding behavior is deprecated, and will be changed in a future version of Qt.

But in Qt6 it asserts:

QFATAL : TestKoColorSet::testLoadACT() ASSERT: "n <= d.size - pos" in file /usr/include/x86_64-linux-gnu/qt6/QtCore/qbytearray.h, line 571
FAIL!  : TestKoColorSet::testLoadACT() Received a fatal error.
   Loc: [/usr/include/x86_64-linux-gnu/qt6/QtCore/qbytearray.h(571)]

And our test file is 2 bytes short of divisible by 3. So, let's
make sure that we only load bytes that exist