[qt6] Fix linking issues with forwarding methods of the new QList
tldr; if you declare an exportable constant in a header, then declare
it as constexpr, not as a usual const. It will make it properly
accessible in expressions like {}.

The new version of QList in Qt6 uses move-construction and
(not-very-)perfect forwarding in QList<T>::append() method.
It means that in the construction like:

QList<QPair<int,int>> list;
list.append({KoColorSpaceMathsTraits<quint16>::min,
             KoColorSpaceMathsTraits<quint16>::max});

the code will pass a non-const-reference to the temporary
object, which, in turn, will try to aquire an address of
KoColorSpaceMathsTraits<quint16>::min.

The problem is that KoColorSpaceMathsTraits<quint16>::min is
placed in a different shared library and we have all the inlines
hidden by default in CMake settings using the follwoing options
(inside ECM):

set(CMAKE_CXX_VISIBILITY_PRESET hidden)
set(CMAKE_VISIBILITY_INLINES_HIDDEN 1)

Which means that the object will **NOT** be found by the linker.
To resolve the issue we should declare the inlines as `constexpr`

The rule of thumb: if you have an **exportable** constant declared in
a header file, declare it constexpr instead of const.

PS:
std::vector has the same issue when faced with the hidden inlined
variable and an on-the-fly type conversion of an std::pair.