Fix default values for gaps in Knife Tool
Before this commit, all spinboxes would hold a default value of 100,
had a minimum value of 0, the Thin and Thick labels were switched,
and the Thick value wasn't selected by default.
This commit adds a sane value for every type of gap for a A4 300dpi
image (considered default in Krita), changes the minimum value to 0
(in case someone needs it - it still works correctly) and the Thick type
is selected by default.