Add snapping to the Knife Tool
Before this commit, there was no way to make a perfectly
horizontal or vertical line. This commit adds snapping in two
cases: 1) if the Shift button is pressed, or 2) if the angle
of the line is very close to the "nice" angle (which is an angle
that is divisible by 15 degrees).