Try to optimize Knife Tool for a big number of shapes
Before this commit, Knife Tool would cut every selected shape,
no matter where it was located. This commit adds a check whether a shape
has a chance being crossed by the gutter, and if not, it skips
calculations and just adds the shape to the resulting collection of shapes.

This was done in order to make the CutThroughShapeStrategy faster in
case of huge numbers of shapes. It's not clear if it helped in a
significant way.