Implement "relayoutIsBlocked" and "fontMatchingDisabled" to speed up glyph palette.
This the former makes it so relayout does not happen automatically, while
the second disables font charmap matching, which means that Krita will
only use the first matched font from the FFWWS converter. In the case
of the glyph palette, this is always going to be the correct font, and
it can speed up font handling greatly.