Change comparing with 0 to qFuzzyIsNull in algebra functions
Before this commit, the code either used qFuzzyCompare(x, 0) or if (x == 0) to compare the value with 0 despite x being a floating point number. This commit changes it to qFuzzyIsNull.

(Note that according to Qt docs, comparing with 0.0 using qFuzzyCompare will not work, so it's not just making it more numerically stable, but fixing a known issue).