First version of remove gutter in the Knife Tool
Before this commit, there was no way to remove a gutter from
between vector shapes that was added with the Knife Tool
(or just manually). The user would have to recreate the shape
manually if they couldn't undo anymore.
This commit adds a feature to remove a gutter automatically.
The feature doesn't work fully correctly yet, but it should handle
the most basic of cases.