Fix issues with calculating a gutter shape for Knife Tool
Before this commit, the middleLine in the removeGutterOneEnd
function was incorrectly calculated, which was breaking
the rest of the calculations. Moreover if the intersection point
wasn't useful because it was too far away, the function
returned a simple result instead of the good one that would happen in
case of no intersection.

This commit fixes both of those issues. The end result might still be
incorrect due to two known issues:
1) in case of the intersection point being close to the middle line
but still far away from the situation,
2) because of the broken uniting of all three shapes.