[win] Implement Tablet API switching using Qt6 interfaces
Qt6 has a built-in interface for switching the tablet API. Not only it
can now switch WinTab/WinInk, it can also do that on the fly, without
restarting the application.

The patch also removes the custom implementation of WinTab and WinInk
protocol in Krita using event filters. We have not used these files for more than
5 years, so noone really knows/cares if they do actually work.