Allow tool to remember opacity.
Currently, opacity is saved only in brush presets. When a tool that is
not the Freehand Brush is active, its opacity is lost when the active
brush is switched. This change updates the tool opacity implementation
to override the brush preset opacity when the Freehand Brush is not
active. The preexisting functionality of saving each brush's opacity
individually is retained.