Qt6: Fix crash with certain brush presets
Presets with a setting using only Fuzzy Dab/Stroke or Drawing Angle
with the curve mode set to Max, Min, or Difference would crash.

These sensors aren't added to the sensorValues list in
KisCurveOption::computeValueComponents, and the code didn't
check if the list was empty before using it. Those curve modes
pass the list to std::max/min_element, which now crash
with an empty list. With Qt5 they returned fuzzy null.

(Shoutout to the brush 'wojtryb6 N 18 complementary'
for triggering this obscure bug.)