Qt6: Fix brush editor height issue
Don't set the preset settings widget's minimumSize to its sizeHint.
It's not needed, and as it now hints the full height of the settings
list, switching to a MyPaint brush caused the widget to go offscreen.

Also fix this in attached mode by not setting the popup button
size policy to Fixed. This also lets the size be kept between
detached and attached mode.