Improve KisAlgebra2D::getParallelLines
Before this commit, KisAlgebra2D::getParallelLines function
was long and complicated, and gave relatively ugly results
(the lines always started at either the same x or the same y).
This commit changes the implementation to a much simpler,
shorter, and generally neater version, that returns lines
that are the exact same length as the original line and have start
and end points perfectly symmetrically located at both sides of the
original start and end point.