Make Remove Gutter mode work on unselected shapes too
Before this commit, Remove Gutter mode in the Knife Tool could
only work on selected shapes. This commit ensures it can work on all
shapes and that it preserves the selectedness of the shapes
(shapes that were selected, will stay selected, and if any shape that
was involved in the removing was selected, the result would also be
selected; it should also work correctly with an undo).