Make Add Gutter work on unselected shapes too
Before this commit, Add Gutter mode in Knife Tool would only
work on selected shapes. This commit ensures it can work on all
shapes on a layer.
It should preserve the selectedness of the shapes, meaning that
if a shape is being cut through, the result should be selected or
unselected depending on whether the original shape was selected or not.
It should also work correctly with an undo.