Cut only through the shapes the knife line goes through
Before this commit, the Add Gutter mode in the Knife Tool
would cut through all shapes. This commit ensures that only
the shapes that were on the gutter are cut.

To be cut, a shape must either contain one of the end points of the
cutting gutter, or contain a part of the cutting gutter.

Note: the rounded end of the gutter is just a cosmetic, and
that part doesn't cut through anything. What matters are the ending
points and the two straight lines forming the gutter.
It should still wokr intuitively enough.